refactor(config): split runtime and validator helpers
